Clean up the FormField and update snapshots
