+++ packages/components/src/FormField/FormField.test.tsx
@@ -47,13 +47,13 @@ it("renders correctly with a placeholder", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440003"
+        htmlFor="123e4567-e89b-12d3-a456-426655440002"
       >
         My placeholder
       </label>
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440003"
+        id="123e4567-e89b-12d3-a456-426655440002"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -77,11 +77,11 @@ it("renders correctly as small", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440005"
+        htmlFor="123e4567-e89b-12d3-a456-426655440003"
       />
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440005"
+        id="123e4567-e89b-12d3-a456-426655440003"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -105,11 +105,11 @@ it("renders correctly in a readonly state", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440007"
+        htmlFor="123e4567-e89b-12d3-a456-426655440004"
       />
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440007"
+        id="123e4567-e89b-12d3-a456-426655440004"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -134,12 +134,12 @@ it("renders correctly in a disabled state", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440009"
+        htmlFor="123e4567-e89b-12d3-a456-426655440005"
       />
       <input
         className="formField"
         disabled={true}
-        id="123e4567-e89b-12d3-a456-426655440009"
+        id="123e4567-e89b-12d3-a456-426655440005"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -163,11 +163,11 @@ it("renders a field with error", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440011"
+        htmlFor="123e4567-e89b-12d3-a456-426655440006"
       />
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440011"
+        id="123e4567-e89b-12d3-a456-426655440006"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
