+++ packages/components/src/FormField/FormSpinner.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import styles from "./FormField.css";
+import { Spinner } from "../Spinner";
+
+export function FormSpinner() {
+  return (
+    <span className={styles.postfix}>
+      <Spinner size="small" />
+    </span>
+  );
+}
