+++ packages/components/src/FormField/index.ts
@@ -1,1 +1,7 @@
 export * from "./FormField";
+
+export {
+  FormFieldProps,
+  FormFieldTypes,
+  FieldActionsRef,
+} from "./FormFieldTypes";
