+++ packages/components/src/InputNumber/InputNumber.test.tsx
@@ -23,7 +23,7 @@ it("renders an input type number", () => {
       <input
         className="formField"
         id="123e4567-e89b-12d3-a456-426655440001"
-        name="generatedName--123e4567-e89b-12d3-a456-426655440002"
+        name="generatedName--123e4567-e89b-12d3-a456-426655440001"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
