+++ packages/components/src/InputText/InputText.test.tsx
@@ -54,13 +54,13 @@ it("renders a textarea", () => {
     >
       <label
         className="label textareaLabel"
-        htmlFor="123e4567-e89b-12d3-a456-426655440003"
+        htmlFor="123e4567-e89b-12d3-a456-426655440002"
       >
         Describe your favourite colour?
       </label>
       <textarea
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440003"
+        id="123e4567-e89b-12d3-a456-426655440002"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -93,13 +93,13 @@ it("renders a textarea with 4 rows", () => {
     >
       <label
         className="label textareaLabel"
-        htmlFor="123e4567-e89b-12d3-a456-426655440005"
+        htmlFor="123e4567-e89b-12d3-a456-426655440003"
       >
         Describe your favourite colour?
       </label>
       <textarea
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440005"
+        id="123e4567-e89b-12d3-a456-426655440003"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
