+++ packages/components/src/InputTime/InputTime.test.tsx
@@ -50,11 +50,11 @@ it("renders an initial time when given 'defaultValue'", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440003"
+        htmlFor="123e4567-e89b-12d3-a456-426655440002"
       />
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440003"
+        id="123e4567-e89b-12d3-a456-426655440002"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -81,11 +81,11 @@ it("renders correctly in a readonly state", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440005"
+        htmlFor="123e4567-e89b-12d3-a456-426655440003"
       />
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440005"
+        id="123e4567-e89b-12d3-a456-426655440003"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -113,11 +113,11 @@ it("adds a error border when invalid", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440007"
+        htmlFor="123e4567-e89b-12d3-a456-426655440004"
       />
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440007"
+        id="123e4567-e89b-12d3-a456-426655440004"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
@@ -143,11 +143,11 @@ it("should set the value when given 'value' and 'onChange'", () => {
     >
       <label
         className="label"
-        htmlFor="123e4567-e89b-12d3-a456-426655440009"
+        htmlFor="123e4567-e89b-12d3-a456-426655440005"
       />
       <input
         className="formField"
-        id="123e4567-e89b-12d3-a456-426655440009"
+        id="123e4567-e89b-12d3-a456-426655440005"
         onBlur={[Function]}
         onChange={[Function]}
         onFocus={[Function]}
